+++ examples/room_ism_from_rt60.py
@@ -1,0 +1,66 @@
+"""
+This example creates a room with reverberation time specified by inverting Sabine's formula.
+
+This results in a reverberation time slightly longer than desired.
+"""
+import matplotlib.pyplot as plt
+import numpy as np
+from scipy.io import wavfile
+
+import pyroomacoustics as pra
+
+if __name__ == "__main__":
+
+    # The desired reverberation time and dimensions of the room
+    rt60_tgt = 0.3  # seconds
+    room_dim = [10, 7.5, 3.5]  # meters
+
+    # import a mono wavfile as the source signal
+    # the sampling frequency should match that of the room
+    fs, audio = wavfile.read("examples/input_samples/cmu_arctic_us_aew_a0001.wav")
+
+    # We invert Sabine's formula to obtain the parameters for the ISM simulator
+    e_absorption, max_order = pra.inverse_sabine(rt60_tgt, room_dim)
+
+    # Create the room
+    room = pra.ShoeBox(
+        room_dim, fs=fs, materials=pra.Material(e_absorption), max_order=max_order
+    )
+
+    # place the source in the room
+    room.add_source([2.5, 3.73, 1.76], signal=audio, delay=1.3)
+
+    # define the locations of the microphones
+    mic_locs = np.c_[
+        [6.3, 4.87, 1.2], [6.3, 4.93, 1.2],  # mic 1  # mic 2
+    ]
+
+    # finally place the array in the room
+    room.add_microphone_array(mic_locs)
+
+    # Run the simulation (this will also build the RIR automatically)
+    room.simulate()
+
+    # measure the reverberation time
+    rt60 = room.measure_rt60()
+    print("The desired RT60 was {}".format(rt60_tgt))
+    print("The measured RT60 is {}".format(rt60[1, 0]))
+
+    # Create a plot
+    plt.figure()
+
+    # plot one of the RIR. both can also be plotted using room.plot_rir()
+    rir_1_0 = room.rir[1][0]
+    plt.subplot(2, 1, 1)
+    plt.plot(np.arange(len(rir_1_0)) / room.fs, rir_1_0)
+    plt.title("The RIR from source 0 to mic 1")
+    plt.xlabel("Time [s]")
+
+    # plot signal at microphone 1
+    plt.subplot(2, 1, 2)
+    plt.plot(room.mic_array.signals[1, :])
+    plt.title("Microphone 1 signal")
+    plt.xlabel("Time [s]")
+
+    plt.tight_layout()
+    plt.show()